Avoid socket answer when user completion doesn't match
While the user is writing a username that isn't matched for autocompletion,
as an error is raised on the browser when an empty array is returned,
the server simply does not return anything. It saves on socket requests and
leaves the browser-side deal with it.
It fixes the error with an undefined array